Escape from circular dependency hell
